# artifacts/api-server/src/routes/wallet.ts
import { Router } from "express";
import { db } from "@workspace/db";
import { walletsTable, transactionsTable, notificationsTable } from "@workspace/db";
import { eq, sql, and, gte } from "drizzle-orm";
import { requireAuth } from "../middlewares/auth";

const router = Router();

router.get("/", requireAuth, async (req, res) => {
  try {
    const [wallet] = await db.select().from(walletsTable).where(eq(walletsTable.userId, req.userId!));

    if (!wallet) {
      res.status(404).json({ error: "Wallet not found" });
      return;
    }

    res.json({
      id: wallet.id,
      userId: wallet.userId,
      balance: parseFloat(wallet.balance),
      pendingBalance: parseFloat(wallet.pendingBalance),
      currency: wallet.currency,
      updatedAt: wallet.updatedAt,
    });
  } catch (e) {
    req.log.error({ e }, "Error getting wallet");
    res.status(500).json({ error: "Failed to get wallet" });
  }
});

router.get("/portfolio", requireAuth, async (req, res) => {
  try {
    const userId = req.userId!;
    const [wallet] = await db.select().from(walletsTable).where(eq(walletsTable.userId, userId));

    const now = new Date();
    const startOfMonth = new Date(now.getFullYear(), now.getMonth(), 1);

    const transactions = await db.select().from(transactionsTable)
      .where(
        and(
          eq(transactionsTable.status, "completed"),
          gte(transactionsTable.createdAt, startOfMonth)
        )
      );

    let monthlyIncome = 0;
    let monthlySpending = 0;

    for (const tx of transactions) {
      const amount = parseFloat(tx.amount);
      if (tx.receiverId === userId && (tx.type === "transfer" || tx.type === "deposit" || tx.type === "admin_fund")) {
        monthlyIncome += amount;
      }
      if (tx.senderId === userId && (tx.type === "transfer" || tx.type === "withdrawal")) {
        monthlySpending += amount;
      }
    }

    const balance = parseFloat(wallet?.balance || "10000");
    const savingsRate = monthlyIncome > 0 ? Math.round(((monthlyIncome - monthlySpending) / monthlyIncome) * 100) : 42;

    const balanceHistory = [];
    for (let i = 5; i >= 0; i--) {
      const d = new Date();
      d.setMonth(d.getMonth() - i);
      balanceHistory.push({
        date: d.toLocaleDateString("en-US", { month: "short", year: "2-digit" }),
        balance: balance * (0.75 + (5 - i) * 0.05 + Math.random() * 0.05),
      });
    }
    balanceHistory.push({ date: "Now", balance });

    const spendingCategories = [
      { name: "Transfers", amount: monthlySpending * 0.4, color: "#10b981" },
      { name: "Withdrawals", amount: monthlySpending * 0.3, color: "#3b82f6" },
      { name: "Fees", amount: monthlySpending * 0.1, color: "#8b5cf6" },
      { name: "Other", amount: monthlySpending * 0.2, color: "#f59e0b" },
    ];

    const monthlyActivity = [];
    for (let i = 5; i >= 0; i--) {
      const d = new Date();
      d.setMonth(d.getMonth() - i);
      monthlyActivity.push({
        month: d.toLocaleDateString("en-US", { month: "short" }),
        income: i === 0 ? monthlyIncome : (200 + Math.random() * 800),
        spending: i === 0 ? monthlySpending : (100 + Math.random() * 400),
      });
    }

    res.json({
      netWorth: balance,
      monthlyIncome: monthlyIncome || 3200,
      monthlySpending: monthlySpending || 1850,
      savingsRate: Math.min(Math.max(savingsRate, 0), 100),
      balanceHistory,
      spendingCategories,
      monthlyActivity,
    });
  } catch (e) {
    req.log.error({ e }, "Error getting portfolio");
    res.status(500).json({ error: "Failed to get portfolio" });
  }
});

router.post("/deposit", requireAuth, async (req, res) => {
  const { amount } = req.body;

  if (!amount || amount <= 0) {
    res.status(400).json({ error: "Valid amount is required" });
    return;
  }

  if (amount > 50000) {
    res.status(400).json({ error: "Maximum deposit amount is $50,000" });
    return;
  }

  try {
    const userId = req.userId!;

    await db.update(walletsTable)
      .set({ balance: sql`${walletsTable.balance} + ${amount}`, updatedAt: new Date() })
      .where(eq(walletsTable.userId, userId));

    const [tx] = await db.insert(transactionsTable).values({
      type: "deposit",
      amount: amount.toString(),
      status: "completed",
      receiverId: userId,
      note: "Account deposit",
    }).returning();

    await db.insert(notificationsTable).values({
      userId,
      type: "transaction",
      title: "Deposit Successful",
      message: `$${parseFloat(amount).toFixed(2)} has been added to your account.`,
    });

    const [updatedWallet] = await db.select().from(walletsTable).where(eq(walletsTable.userId, userId));

    res.json({
      transaction: {
        id: tx.id,
        type: tx.type,
        amount: parseFloat(tx.amount),
        status: tx.status,
        note: tx.note,
        receiverId: tx.receiverId,
        createdAt: tx.createdAt,
        updatedAt: tx.updatedAt,
      },
      message: "Deposit processed successfully",
    });
  } catch (e) {
    req.log.error({ e }, "Error processing deposit");
    res.status(500).json({ error: "Failed to process deposit" });
  }
});

export default router;