# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sterling Crest Bank API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface RegisterBody {
  email: string;
  username: string;
  phone: string;
  country: string;
  password: string;
  confirmPassword: string;
}

export interface VerifyOtpBody {
  email: string;
  otp: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  password: string;
  confirmPassword: string;
}

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  active: "active",
  frozen: "frozen",
  pending_verification: "pending_verification",
} as const;

export interface User {
  id: number;
  email: string;
  username: string;
  phone?: string;
  country?: string;
  status: UserStatus;
  kycLevel: number;
  hasPin: boolean;
  createdAt: string;
  accountNumber?: string;
}

export interface AuthResponse {
  user: User;
  token?: string;
  requiresOtp?: boolean;
  message?: string;
}

export interface Wallet {
  id: number;
  userId: number;
  balance: number;
  pendingBalance: number;
  currency: string;
  updatedAt?: string;
}

export interface UserWithWallet {
  id: number;
  email: string;
  username: string;
  phone?: string;
  country?: string;
  status: string;
  kycLevel: number;
  hasPin: boolean;
  createdAt: string;
  accountNumber?: string;
  wallet?: Wallet;
}

export interface UserProfile {
  id: number;
  email: string;
  username: string;
  phone?: string;
  country?: string;
  status: string;
  kycLevel: number;
  hasPin: boolean;
  accountNumber: string;
  createdAt: string;
  wallet?: Wallet;
}

export interface SetPinBody {
  pin: string;
  confirmPin: string;
}

export interface VerifyPinBody {
  pin: string;
}

export interface BalancePoint {
  date: string;
  balance: number;
}

export interface SpendingCategory {
  name: string;
  amount: number;
  color?: string;
}

export interface MonthlyActivity {
  month: string;
  income: number;
  spending: number;
}

export interface Portfolio {
  netWorth: number;
  monthlyIncome: number;
  monthlySpending: number;
  savingsRate: number;
  balanceHistory: BalancePoint[];
  spendingCategories: SpendingCategory[];
  monthlyActivity: MonthlyActivity[];
}

export interface DepositBody {
  amount: number;
}

export interface TransferBody {
  recipientAccountNumber: string;
  amount: number;
  pin: string;
  note?: string;
}

export type WithdrawBodyMethod =
  (typeof WithdrawBodyMethod)[keyof typeof WithdrawBodyMethod];

export const WithdrawBodyMethod = {
  cash_app: "cash_app",
  paypal: "paypal",
  revolut: "revolut",
  bank_transfer: "bank_transfer",
  bank_card: "bank_card",
} as const;

export interface WithdrawBody {
  amount: number;
  method: WithdrawBodyMethod;
  destination: string;
  bankName?: string;
  accountDetails?: string;
  pin: string;
  note?: string;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  transfer: "transfer",
  withdrawal: "withdrawal",
  deposit: "deposit",
  gift_card: "gift_card",
  admin_fund: "admin_fund",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  status: TransactionStatus;
  note?: string;
  declineReason?: string;
  senderId?: number;
  receiverId?: number;
  senderUsername?: string;
  receiverUsername?: string;
  method?: string;
  destination?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface TransactionResponse {
  transaction: Transaction;
  message?: string;
}

export interface TransactionListResponse {
  transactions: Transaction[];
  total: number;
}

export interface TransactionReceipt {
  transaction: Transaction;
  bank: string;
  generatedAt: string;
}

export type KycStatusStatus =
  (typeof KycStatusStatus)[keyof typeof KycStatusStatus];

export const KycStatusStatus = {
  not_started: "not_started",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface KycStatus {
  level: number;
  status: KycStatusStatus;
  rejectionReason?: string;
  submittedAt?: string;
  reviewedAt?: string;
}

export type KycSubmitBodyIdType =
  (typeof KycSubmitBodyIdType)[keyof typeof KycSubmitBodyIdType];

export const KycSubmitBodyIdType = {
  passport: "passport",
  drivers_license: "drivers_license",
  national_id: "national_id",
} as const;

export interface KycSubmitBody {
  fullName: string;
  dateOfBirth: string;
  address: string;
  idType: KycSubmitBodyIdType;
  idNumber: string;
  idFrontImage: string;
  idBackImage?: string;
  selfieImage?: string;
}

export interface BankCard {
  id: number;
  cardHolderName: string;
  last4: string;
  brand: string;
  expiryMonth: string;
  expiryYear: string;
  createdAt: string;
}

export interface AddCardBody {
  cardNumber: string;
  cardHolderName: string;
  expiryMonth: string;
  expiryYear: string;
  cvv: string;
}

export interface GiftCardBody {
  brand: string;
  code: string;
  amount: number;
  frontImage: string;
  backImage?: string;
  pin: string;
}

export interface Notification {
  id: number;
  type: string;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface SupportMessageBody {
  message: string;
}

export interface SupportMessage {
  id: number;
  message: string;
  isFromUser: boolean;
  createdAt: string;
}

export interface RegionalBank {
  name: string;
  logo?: string;
  code: string;
}

export interface IpLocation {
  country: string;
  countryCode: string;
  region?: string;
  city?: string;
  banks: RegionalBank[];
}

export type GetTransactionsParams = {
  type?: GetTransactionsType;
  status?: GetTransactionsStatus;
  limit?: number;
  offset?: number;
};

export type GetTransactionsType =
  (typeof GetTransactionsType)[keyof typeof GetTransactionsType];

export const GetTransactionsType = {
  all: "all",
  transfer: "transfer",
  withdrawal: "withdrawal",
  deposit: "deposit",
  gift_card: "gift_card",
  admin_fund: "admin_fund",
} as const;

export type GetTransactionsStatus =
  (typeof GetTransactionsStatus)[keyof typeof GetTransactionsStatus];

export const GetTransactionsStatus = {
  all: "all",
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export type TelegramWebhookBody = { [key: string]: unknown };

export type SupportTelegramWebhookBody = { [key: string]: unknown };