# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sterling Crest Bank API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/set-pin:
    post:
      operationId: setPin
      tags: [users]
      summary: Set transaction PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPinBody"
      responses:
        "200":
          description: PIN set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /users/verify-pin:
    post:
      operationId: verifyPin
      tags: [users]
      summary: Verify transaction PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyPinBody"
      responses:
        "200":
          description: PIN verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /wallet/deposit:
    post:
      operationId: deposit
      tags: [wallet]
      summary: Deposit funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositBody"
      responses:
        "200":
          description: Deposit processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get transaction history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [all, transfer, withdrawal, deposit, gift_card, admin_fund]
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, completed, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /transactions/transfer:
    post:
      operationId: transfer
      tags: [transactions]
      summary: Transfer funds to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferBody"
      responses:
        "201":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Insufficient funds or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/withdraw:
    post:
      operationId: withdraw
      tags: [transactions]
      summary: Withdraw funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawBody"
      responses:
        "201":
          description: Withdrawal request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transactions/{id}/receipt:
    get:
      operationId: getTransactionReceipt
      tags: [transactions]
      summary: Get transaction receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Receipt data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionReceipt"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    SetPinBody:
      type: object
      properties:
        pin:
          type: string
        confirmPin:
          type: string
      required: [pin, confirmPin]
    VerifyPinBody:
      type: object
      properties:
        pin:
          type: string
      required: [pin]
...
    DepositBody:
      type: object
      properties:
        amount:
          type: number
      required: [amount]
    TransferBody:
      type: object
      properties:
        recipientAccountNumber:
          type: string
        amount:
          type: number
        pin:
          type: string
        note:
          type: string
      required: [recipientAccountNumber, amount, pin]
    WithdrawBody:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [cash_app, paypal, revolut, bank_transfer, bank_card]
        destination:
          type: string
        bankName:
          type: string
        accountDetails:
          type: string
        pin:
          type: string
        note:
          type: string
      required: [amount, method, destination, pin]
    TransactionResponse:
      type: object
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
        message:
          type: string
      required: [transaction]
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [transfer, withdrawal, deposit, gift_card, admin_fund]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed]
        note:
          type: string
        declineReason:
          type: string
        senderId:
          type: integer
        receiverId:
          type: integer
        senderUsername:
          type: string
        receiverUsername:
          type: string
        method:
          type: string
        destination:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, type, amount, status, createdAt]
    TransactionListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
      required: [transactions, total]
    TransactionReceipt:
      type: object
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
        bank:
          type: string
        generatedAt:
          type: string
          format: date-time
      required: [transaction, bank, generatedAt]
...